Refactor functions for chars

+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToBooleanFunction.java
@@ -3,16 +3,15 @@ package ar.com.kfgodel.function.intertype.bytes;
 import java.util.function.Function;
 
 /**
- * This type defines an operator for primitive boolean types
  * Created by kfgodel on 23/07/17.
  */
-public interface ByteToBooleanFunction extends Function<Byte, Boolean> {
+public interface ByteToBooleanFunction extends Function<Integer, Boolean> {
 
   boolean apply(byte aBoolean);
 
   @Override
-  default Boolean apply(Byte aBoolean){
-    return this.apply(aBoolean.byteValue());
+  default Boolean apply(Integer aBoolean){
+    return this.apply(aBoolean.intValue());
   }
 
 }
